Search for expired entry before removing from index
This enhances the cleanup of orphaned sessions in the principal index, by first getting all the keys matching the pattern spring:session:expirations:*, where the last part is the timestamp rounded up to the nearest minute. Then it filters all keys with timestamp < last minute and checks if the key inside the Set exists before it removes it from the principal index.

This applies the algorithm described at https://github.com/spring-projects/spring-session/pull/2869#issuecomment-2016136675.